backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.6.1
chromadb==0.4.22
langchain==0.0.340
langchain-community==0.0.10
tiktoken==0.5.2
sse-starlette==1.6.5
pypdf==3.17.4
httpx==0.25.2
